requirements-dev.txt
pytest>=8.0
ruff>=0.9.0
pyyaml>=6.0
